+++ app/classes/Framadate/Services/AdminPollService.php
@@ -117,7 +117,10 @@ class AdminPollService {
 
         $slots = $this->pollService->allSlotsByPoll($poll);
 
-        if (count($slots) === 1) {
+        // We can't delete the last slot
+        if ($poll->format == 'D' && count($slots) === 1 && strpos($slots[0]->moments, ',') === -1) {
+            return false;
+        } elseif ($poll->format == 'A' && count($slots) === 1) {
             return false;
         }
 
@@ -196,6 +199,8 @@ class AdminPollService {
      * @return bool true if added
      */
     public function addDateSlot($poll_id, $datetime, $new_moment) {
+        $this->logService->log('ADD_SLOT', 'id:' . $poll_id . ', datetime:' . $datetime . ', moment:' . $new_moment);
+
         $slots = $this->slotRepository->listByPollId($poll_id);
         $result = $this->findInsertPosition($slots, $datetime, $new_moment);
 
@@ -243,6 +248,8 @@ class AdminPollService {
      * @return bool true if added
      */
     public function addClassicSlot($poll_id, $title) {
+        $this->logService->log('ADD_SLOT', 'id:' . $poll_id . ', title:' . $title);
+
         $slots = $this->slotRepository->listByPollId($poll_id);
 
         // Check if slot already exists